+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0.2
+
+* Fixed the inability to remove the default sorts.
+* Added a test to verify the default sorts behavior.
+
 ## 1.0.1
 
 * Fixed formatting issues